Add support for the large majority of WHEA events
Details of all known WHEA events and their support status has been
added to the README. Additionally, I've added test cases for all WHEA
events for which an event is available on my own system. This is a small
minority of the total number of event types, so we really need to source
further sample events from the community.